requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.124.4
greenlet==3.1.1
h11==0.16.0
idna==3.11
asyncpg
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==0.44.0
typing_extensions==4.13.2
uvicorn==0.33.0
